git-local path variable check added

+++ spec/Include/IncludeRemote.spec.js
@@ -10,6 +10,10 @@ const Log = require('log');
 const path = require('path');
 const fs = require('fs');
 
+const LOCAL_REPO_NOT_DEFINED_MESSAGE = "SPEC_GIT_LOCAL_REPO_PATH is not defined. Test will be skipped";
+const NO_ROOT_PERMISSION_MESSAGE = "No root permission. Test will be skipped";
+const WINDOWS_SPECIFIED_TEST_MESSAGE = "Windows platform specified test will be skipped";
+
 describe('Remote relative option is enabled - ', () => {
 
   let builder;
@@ -56,10 +60,10 @@ describe('Remote relative option is enabled - ', () => {
           expect(output).toContain('// y.nut (case y abs)\n');
         }
         catch (err) {
-          console.log("No root permission. Test will be skipped");
+          console.log(NO_ROOT_PERMISSION_MESSAGE);
         }
       } else {
-        console.log("Windows platform specified test will be skipped");
+        console.log(WINDOWS_SPECIFIED_TEST_MESSAGE);
       }
     });
   });
@@ -87,13 +91,21 @@ describe('Remote relative option is enabled - ', () => {
   describe('X path by local repo - ', () => {
 
     it('should search Y file in remote repository', () => {
-      let output = builder.machine.execute(`@include "git-local:${process.env.SPEC_GIT_LOCAL_REPO_PATH}/spec/fixtures/include/sample-2/LibA/dirX/x_case_y_github.nut"`);
-      expect(output).toContain('// y.nut (case y remote)\n');
+      if (process.env.SPEC_GIT_LOCAL_REPO_PATH != undefined) {
+        let output = builder.machine.execute(`@include "git-local:${process.env.SPEC_GIT_LOCAL_REPO_PATH}/spec/fixtures/include/sample-2/LibA/dirX/x_case_y_github.nut"`);
+        expect(output).toContain('// y.nut (case y remote)\n');
+      } else {
+        console.log(LOCAL_REPO_NOT_DEFINED_MESSAGE);
+      }
     });
 
     it('should search Y file in remote repository + Y path', () => {
-      let output = builder.machine.execute(`@include "git-local:${process.env.SPEC_GIT_LOCAL_REPO_PATH}/spec/fixtures/include/sample-2/LibA/dirX/x_case_y_rel_local.nut"`);
-      expect(output).toContain('// y.nut (case y rel)\n');
+      if (process.env.SPEC_GIT_LOCAL_REPO_PATH != undefined) {
+        let output = builder.machine.execute(`@include "git-local:${process.env.SPEC_GIT_LOCAL_REPO_PATH}/spec/fixtures/include/sample-2/LibA/dirX/x_case_y_rel_local.nut"`);
+        expect(output).toContain('// y.nut (case y rel)\n');
+      } else {
+        console.log(LOCAL_REPO_NOT_DEFINED_MESSAGE);
+      }
     });
   });
 });
@@ -132,7 +144,7 @@ describe('Remote relative option is not enabled - ', () => {
         expect(output).toContain('// y.nut (case y abs)\n');
       }
       catch (err) {
-        console.log("No root permission. Test will be skipped");
+        console.log(NO_ROOT_PERMISSION_MESSAGE);
       }
     });
   });
@@ -150,26 +162,34 @@ describe('Remote relative option is not enabled - ', () => {
   describe('X path by local repo - ', () => {
 
     it('should search Y file in remote repository', () => {
-      let output = builder.machine.execute(`@include "git-local:${process.env.SPEC_GIT_LOCAL_REPO_PATH}/spec/fixtures/include/sample-2/LibA/dirX/x_case_y_https.nut"`);
-      expect(output).toContain('// y.nut (case y remote)\n');
+      if (process.env.SPEC_GIT_LOCAL_REPO_PATH != undefined) {
+        let output = builder.machine.execute(`@include "git-local:${process.env.SPEC_GIT_LOCAL_REPO_PATH}/spec/fixtures/include/sample-2/LibA/dirX/x_case_y_https.nut"`);
+        expect(output).toContain('// y.nut (case y remote)\n');
+      } else {
+        console.log(LOCAL_REPO_NOT_DEFINED_MESSAGE);
+      }
     });
 
     it('should search Y file by local abs path', () => {
-      try {
-        fs.accessSync("/", fs.constants.W_OK);
-        if (fs.existsSync("/dirC")) {
+      if (process.env.SPEC_GIT_LOCAL_REPO_PATH != undefined) {
+        try {
+          fs.accessSync("/", fs.constants.W_OK);
+          if (fs.existsSync("/dirC")) {
+            fs.unlinkSync("/dirC/y.nut");
+            fs.rmdirSync("/dirC", { recursive: true });
+          }
+          fs.mkdirSync("/dirC");
+          fs.writeFileSync("/dirC/y.nut", "// y.nut (case y abs)\n");
+          let output = builder.machine.execute(`@include "git-local:${process.env.SPEC_GIT_LOCAL_REPO_PATH}/spec/fixtures/include/sample-2/LibA/dirX/x_case_y_abs_local_slash2.nut"`);
           fs.unlinkSync("/dirC/y.nut");
           fs.rmdirSync("/dirC", { recursive: true });
+          expect(output).toContain('// y.nut (case y abs)\n');
         }
-        fs.mkdirSync("/dirC");
-        fs.writeFileSync("/dirC/y.nut", "// y.nut (case y abs)\n");
-        let output = builder.machine.execute(`@include "git-local:${process.env.SPEC_GIT_LOCAL_REPO_PATH}/spec/fixtures/include/sample-2/LibA/dirX/x_case_y_abs_local_slash2.nut"`);
-        fs.unlinkSync("/dirC/y.nut");
-        fs.rmdirSync("/dirC", { recursive: true });
-        expect(output).toContain('// y.nut (case y abs)\n');
-      }
-      catch (err) {
-        console.log("No root permission. Test will be skipped");
+        catch (err) {
+          console.log(NO_ROOT_PERMISSION_MESSAGE);
+        }
+      } else {
+        console.log(LOCAL_REPO_NOT_DEFINED_MESSAGE);
       }
     });
   });
@@ -201,7 +221,7 @@ describe('Remote relative option is not enabled - ', () => {
         expect(output).toContain('// y.nut (case y abs)\n');
       }
       catch (err) {
-        console.log("No root permission. Test will be skipped");
+        console.log(NO_ROOT_PERMISSION_MESSAGE);
       }
     });
   });
